appWeb/fastapi/requirements.txt
fastapi
python-jose
passlib
pydantic
sqlalchemy
uvicorn
pymysql
python-multipart
sqlalchemy-utils
bcrypt
cryptography